requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.15.2
Faker==40.1.2
sqlparse==0.5.5
